Add files via upload
This Supply Management system in Java is easy to understand and handle. The System comes with two handlers. One is the admin and the other is the cashier. The admin has full access to the system. The admin can create as many cashiers he wants to. He also can view stocks and products. Moreover, he controls the transaction process in this system. Whereas, the cashiers can add products and can generate bills. Also, they can update and delete them later.